rename InputValues -> QueryArguments
